updated database quantity datatype to DECIMAL

+++ database.sql
@@ -20,10 +20,10 @@ CREATE TABLE "palette" (
 	"hex4" varchar(7) NOT NULL,
 	"hex5" varchar(7) NOT NULL,
 	"img_url" varchar(2083) NOT NULL,
-	"quantity0" int NOT NULL,
-	"quantity1" int NOT NULL,
-	"quantity2" int NOT NULL,
-	"quantity3" int NOT NULL,
-	"quantity4" int NOT NULL,
-	"quantity5" int NOT NULL
+	"quantity0" DECIMAL(5, 2) NOT NULL,
+	"quantity1" DECIMAL(5, 2) NOT NULL,
+	"quantity2" DECIMAL(5, 2) NOT NULL,
+	"quantity3" DECIMAL(5, 2) NOT NULL,
+	"quantity4" DECIMAL(5, 2) NOT NULL,
+	"quantity5" DECIMAL(5, 2) NOT NULL
 	);
